Migrate gdpr command (#353)
* Migrate gdpr command

* Adjust locale key names
