+++ src/main/resources/locale.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=
 command.debug.description=
 command.debug.title=
 command.gdpr.description=
-command.gdpr.sub.delete=
-command.gdpr.sub.delete.received=
-command.gdpr.sub.delete.scheduled=
-command.gdpr.sub.request=
-command.gdpr.sub.request.received=
-command.gdpr.sub.request.requested=
+command.gdpr.delete.description=
+command.gdpr.delete.message.received=
+command.gdpr.delete.message.scheduled=
+command.gdpr.request.description=
+command.gdpr.request.message.received=
+command.gdpr.request.message.requested=
 command.info.art=
 command.info.contributor=
 command.info.description=
