+++ src/main/resources/locale_de_DE.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=Ist der Kanal aktiv?
 command.debug.description=Debug Informationen für den Server
 command.debug.title=Debug für %GUILD%
 command.gdpr.description=Beantrage eine Kopie oder die Löschung deiner Daten.
-command.gdpr.sub.delete=Beantrage die Löschung deiner Daten.
-command.gdpr.sub.delete.received=Deine Daten sind zur Löschung freigegeben. Sie werden innerhalb der nächsten Stunden entfernt.
-command.gdpr.sub.delete.scheduled=Deine Daten sind bereits zur Löschung freigegeben. Sie werden in den nächsten Stunden entfernt.
-command.gdpr.sub.request=Beantrage eine Kopie deiner Daten.
-command.gdpr.sub.request.received=Wir haben deine Anfrage erhalten. Du wirst deine Daten innerhalb der nächsten Stunden erhalten. Du kannst deine Daten alle 30 Tage anfragen. **Du musst einen Server mit dem Bot teilen, um deine Daten zu erhalten. Andernfalls können sie nicht an dich gesendet werden. Stelle außerdem sicher, dass du private Nachrichten von diesem Server empfangen kannst.**
-command.gdpr.sub.request.requested=Du hast deine Daten innerhalb der letzten 30 Tage angefragt. Du kannst deine Daten nur alle 30 Tage anfragen.
+command.gdpr.delete.description=Beantrage die Löschung deiner Daten.
+command.gdpr.delete.message.received=Deine Daten sind zur Löschung freigegeben. Sie werden innerhalb der nächsten Stunden entfernt.
+command.gdpr.delete.message.scheduled=Deine Daten sind bereits zur Löschung freigegeben. Sie werden in den nächsten Stunden entfernt.
+command.gdpr.request.description=Beantrage eine Kopie deiner Daten.
+command.gdpr.request.message.received=Wir haben deine Anfrage erhalten. Du wirst deine Daten innerhalb der nächsten Stunden erhalten. Du kannst deine Daten alle 30 Tage anfragen. **Du musst einen Server mit dem Bot teilen, um deine Daten zu erhalten. Andernfalls können sie nicht an dich gesendet werden. Stelle außerdem sicher, dass du private Nachrichten von diesem Server empfangen kannst.**
+command.gdpr.request.message.requested=Du hast deine Daten innerhalb der letzten 30 Tage angefragt. Du kannst deine Daten nur alle 30 Tage anfragen.
 command.info.art=Grafiken & Branding
 command.info.contributor=Mitwirkende
 command.info.description=Zeigt den Quellcode an!
