+++ src/main/resources/locale_en_US.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=Is Channel active?
 command.debug.description=Debug Information for the guild
 command.debug.title=Debug for %GUILD%
 command.gdpr.description=Request a copy of data or its deletion.
-command.gdpr.sub.delete=Request deletion of your data.
-command.gdpr.sub.delete.received=Your data is scheduled to be deleted. It will be deleted within the next hours.
-command.gdpr.sub.delete.scheduled=Your data is already scheduled to be deleted. It will be deleted within the next hours.
-command.gdpr.sub.request=Request a copy or your data.
-command.gdpr.sub.request.received=We received your request. You will get your data within the next hours. You can request your data every 30 days. **You must share a server with the bot to get your data. Otherwise it cannot be sent to you. Also, make sure that you are enabled to receive private messages from this server.**
-command.gdpr.sub.request.requested=You have requested your data in the last 30 days. You can request your data only every 30 days.
+command.gdpr.delete.description=Request deletion of your data.
+command.gdpr.delete.message.received=Your data is scheduled to be deleted. It will be deleted within the next hours.
+command.gdpr.delete.message.scheduled=Your data is already scheduled to be deleted. It will be deleted within the next hours.
+command.gdpr.request.description=Request a copy or your data.
+command.gdpr.request.message.received=We received your request. You will get your data within the next hours. You can request your data every 30 days. **You must share a server with the bot to get your data. Otherwise it cannot be sent to you. Also, make sure that you are enabled to receive private messages from this server.**
+command.gdpr.request.message.requested=You have requested your data in the last 30 days. You can request your data only every 30 days.
 command.info.art=Artwork & Branding
 command.info.contributor=Contributor
 command.info.description=Get information about me!
