+++ src/main/resources/locale_es_ES.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=¿Está activo el canal?
 command.debug.description=Debug Information for the guild
 command.debug.title=Depuración para %GUILD%
 command.gdpr.description=Solicitar una copia de datos o su eliminación
-command.gdpr.sub.delete=Solicitar el borrado de sus datos.
-command.gdpr.sub.delete.received=Sus datos están programados para ser borrados. Se borrarán en las próximas horas.
-command.gdpr.sub.delete.scheduled=Sus datos ya están programados para ser borrados. Se eliminarán en las próximas horas.
-command.gdpr.sub.request=Solicitar una copia de sus datos.
-command.gdpr.sub.request.received=Hemos recibido su solicitud. Recibirá sus datos en las próximas horas. Puede solicitar sus datos cada 30 días. **Debes compartir un servidor con el bot para obtener tus datos. De lo contrario, no podrá enviártelos. Además, asegúrate de que estás habilitado para recibir mensajes privados de este servidor.**
-command.gdpr.sub.request.requested=Ha solicitado sus datos en los últimos 30 días. Puede solicitar sus datos sólo cada 30 días.
+command.gdpr.delete.description=Solicitar el borrado de sus datos.
+command.gdpr.delete.message.received=Sus datos están programados para ser borrados. Se borrarán en las próximas horas.
+command.gdpr.delete.message.scheduled=Sus datos ya están programados para ser borrados. Se eliminarán en las próximas horas.
+command.gdpr.request.description=Solicitar una copia de sus datos.
+command.gdpr.request.message.received=Hemos recibido su solicitud. Recibirá sus datos en las próximas horas. Puede solicitar sus datos cada 30 días. **Debes compartir un servidor con el bot para obtener tus datos. De lo contrario, no podrá enviártelos. Además, asegúrate de que estás habilitado para recibir mensajes privados de este servidor.**
+command.gdpr.request.message.requested=Ha solicitado sus datos en los últimos 30 días. Puede solicitar sus datos sólo cada 30 días.
 command.info.art=Trabajo artístico y marca
 command.info.contributor=Contribuidor
 command.info.description=¡Obtén información sobre mí!
