+++ src/main/resources/locale_fr_FR.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=La chaîne est-elle activée?
 command.debug.description=Debug Information for the guild
 command.debug.title=Débogage pour %GUILD%
 command.gdpr.description=Requiert une copie des données ou leur suppression.
-command.gdpr.sub.delete=Requête la suppression de tes données.
-command.gdpr.sub.delete.received=Ta suppression de données est programmée. Elles seront supprimées dans les prochaines heures.
-command.gdpr.sub.delete.scheduled=Tes données sont déjà programmées pour être supprimées. Elles seront supprimées dans les prochaines heures.
-command.gdpr.sub.request=Requête une copie de tes données.
-command.gdpr.sub.request.received=Nous avons reçu ta demande. Tu recevras tes données dans les prochaines heures. Tu peux demander tes données tous les 30 jours. **Vous devez partager un serveur avec le bot pour obtenir vos données. Sinon, elles ne pourront pas vous être envoyées. Vérifiez également que vous êtes autorisé à recevoir des messages privés de ce serveur.**
-command.gdpr.sub.request.requested=Tu as demandé tes données au cours des 30 derniers jours. Tu ne peux demander tes données que tous les 30 jours.
+command.gdpr.delete.description=Requête la suppression de tes données.
+command.gdpr.delete.message.received=Ta suppression de données est programmée. Elles seront supprimées dans les prochaines heures.
+command.gdpr.delete.message.scheduled=Tes données sont déjà programmées pour être supprimées. Elles seront supprimées dans les prochaines heures.
+command.gdpr.request.description=Requête une copie de tes données.
+command.gdpr.request.message.received=Nous avons reçu ta demande. Tu recevras tes données dans les prochaines heures. Tu peux demander tes données tous les 30 jours. **Vous devez partager un serveur avec le bot pour obtenir vos données. Sinon, elles ne pourront pas vous être envoyées. Vérifiez également que vous êtes autorisé à recevoir des messages privés de ce serveur.**
+command.gdpr.request.message.requested=Tu as demandé tes données au cours des 30 derniers jours. Tu ne peux demander tes données que tous les 30 jours.
 command.info.art=Travaux d'art et image de marque
 command.info.contributor=Contributeur
 command.info.description=Recevoir des informations sur moi !
