+++ src/main/resources/locale_pt_PT.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=O Canal está activo?
 command.debug.description=Debug Information for the guild
 command.debug.title=Depurar para %GUILD%
 command.gdpr.description=Pedir uma cópia dos dados ou a sua eliminação.
-command.gdpr.sub.delete=Pedir o apagamento dos seus dados.
-command.gdpr.sub.delete.received=Os seus dados estão programados para serem apagados. Os seus dados serão apagados dentro das próximas horas.
-command.gdpr.sub.delete.scheduled=Os seus dados já estão agendados para serem apagados. Será apagado dentro das próximas horas.
-command.gdpr.sub.request=Pedir uma cópia ou os seus dados.
-command.gdpr.sub.request.received=Recebemos o seu pedido. Receberá os seus dados dentro das próximas horas. Pode solicitar os seus dados de30 em 30 dias. **Deve partilhar um servidor com o bot para obter os seus dados. Caso contrário, não poderá ser-lhe enviado. Certifique-se também de que está habilitado a receber mensagens privadas a partir deste servidor.**
-command.gdpr.sub.request.requested=Recebeu os seus dados nos últimos 30 dias. Pode solicitar os seus dados apenas de 30 em 30 dias.
+command.gdpr.delete.description=Pedir o apagamento dos seus dados.
+command.gdpr.delete.message.received=Os seus dados estão programados para serem apagados. Os seus dados serão apagados dentro das próximas horas.
+command.gdpr.delete.message.scheduled=Os seus dados já estão agendados para serem apagados. Será apagado dentro das próximas horas.
+command.gdpr.request.description=Pedir uma cópia ou os seus dados.
+command.gdpr.request.message.received=Recebemos o seu pedido. Receberá os seus dados dentro das próximas horas. Pode solicitar os seus dados de30 em 30 dias. **Deve partilhar um servidor com o bot para obter os seus dados. Caso contrário, não poderá ser-lhe enviado. Certifique-se também de que está habilitado a receber mensagens privadas a partir deste servidor.**
+command.gdpr.request.message.requested=Recebeu os seus dados nos últimos 30 dias. Pode solicitar os seus dados apenas de 30 em 30 dias.
 command.info.art=Obras de arte e Branding
 command.info.contributor=Contribuidor
 command.info.description=Recebe informação sobre mim!
