+++ src/main/resources/locale_ru_RU.properties
@@ -76,12 +76,12 @@ command.debug.channelActive=Активен ли канал?
 command.debug.description=Debug Information for the guild
 command.debug.title=Отладка для %GUILD%
 command.gdpr.description=Запросить копию или удаление ваших данных.
-command.gdpr.sub.delete=Запросить удаление ваших данных.
-command.gdpr.sub.delete.received=Ваши данные запланированы к удалению. Мы удалим их в течении следующих нескольких часов.
-command.gdpr.sub.delete.scheduled=Ваши данные уже запланированы к удалению. Мы удалим их в течении следующих нескольких часов.
-command.gdpr.sub.request=Запросить копию ваших данных.
-command.gdpr.sub.request.received=Мы получили ваш запрос. Вы получите данные в течении следующих нескольких часов. Вы можете запрашивать данные каждые 30 дней. **Ты должен иметь общий сервер с ботом, чтобы получить свои данные. В противном случае они не могут быть отправлены тебе. Также убедись, что у тебя включено получение личных сообщений с этого сервера.**
-command.gdpr.sub.request.requested=Вы запросили ваши данные за последние 30 дней. Вы можете запрашивать данные каждые 30 дней.
+command.gdpr.delete.description=Запросить удаление ваших данных.
+command.gdpr.delete.message.received=Ваши данные запланированы к удалению. Мы удалим их в течении следующих нескольких часов.
+command.gdpr.delete.message.scheduled=Ваши данные уже запланированы к удалению. Мы удалим их в течении следующих нескольких часов.
+command.gdpr.request.description=Запросить копию ваших данных.
+command.gdpr.request.message.received=Мы получили ваш запрос. Вы получите данные в течении следующих нескольких часов. Вы можете запрашивать данные каждые 30 дней. **Ты должен иметь общий сервер с ботом, чтобы получить свои данные. В противном случае они не могут быть отправлены тебе. Также убедись, что у тебя включено получение личных сообщений с этого сервера.**
+command.gdpr.request.message.requested=Вы запросили ваши данные за последние 30 дней. Вы можете запрашивать данные каждые 30 дней.
 command.info.art=Визуальные работы и Брендинг
 command.info.contributor=Помощник
 command.info.description=Информация обо мне!
